Rename yoast reading time hook
